+++ zproject/urls.py
@@ -160,7 +160,7 @@ urlpatterns += patterns('zerver.views',
     url(r'^api/v1/external/pivotal$',       'webhooks.pivotal.api_pivotal_webhook'),
     url(r'^api/v1/external/newrelic$',      'webhooks.newrelic.api_newrelic_webhook'),
     url(r'^api/v1/external/bitbucket$',     'webhooks.bitbucket.api_bitbucket_webhook'),
-    url(r'^api/v1/external/desk$',          'webhooks.api_deskdotcom_webhook'),
+    url(r'^api/v1/external/desk$',          'webhooks.deskdotcom.api_deskdotcom_webhook'),
     url(r'^api/v1/external/stash$',         'webhooks.api_stash_webhook'),
     url(r'^api/v1/external/freshdesk$',     'webhooks.api_freshdesk_webhook'),
     url(r'^api/v1/external/zendesk$',       'webhooks.api_zendesk_webhook'),
